Prohibit plain WIP commits (#12)
Commit messages with just "WIP" may be considered
fine (at the end of the day, they are work in progress), but
when many commits are named the same it's difficult to
review them. It's not much effort to just add a simple number,
so this rule forces the author to do just this.