commitlint.cfg: fix body-prose's false positive
This also has the bonus of removing the brittle mightBeUrl()
function in favour of isValidUrl(). However, doing this exposed
a bug in body-prose caused by a commitlint bug, for which we
already have a workaround in the proper-issue-refs rule, which
we have to use in body-prose as well now.

See:
https://github.com/nblockchain/conventions/pull/28