Favor 'scope' & 'sub-scope' vs 'area' & 'sub-area'
The convention of commitlint upstream documentation is to refer
to the text that goes before the ':' in the title as the
**type(scope)**, but we called it **area(sub-area)** (also
allowed to be written with a slash: **area/sub-area**).

There was no point to really use a different word for it so,
from now on, we will refer to it as scope as well. But we
still don't use type, just scope and sub-scope.