commitlint.config: disable type-empty rule
We disable it because most of the time it didn't work anyway,
due to an upstream commitlint bug[1], and anyway we were using
this rule only as a warning, not an error (because a scope is
not required, e.g. when commit is too broad, that is, affects
almost all possible scopes of the repo).

[1] https://github.com/conventional-changelog/commitlint/issues/3404