scripts/styleChecker.fsx: add
Add styleChecker.fsx script to check style of our F#, TS and
YML codes.

The `git respore package.json` command in the styleChecker.fsx
script was failing with the following error, even when I was
running the `git config --global --add safe.directory '*'`
command in both the styleChecker.fsx script and in the CI. In
the issue [1], it's suggested by someone to use --system instead
of --global in the mentioned command, when the git command is
running in a container, which solved the problem.

```
fatal: detected dubious ownership in repository at '/__w/conventions/conventions'
To add an exception for this directory, call:

	git config --global --add safe.directory /__w/conventions/conventions

Error when running 'git restore package.json'
Fsdk.Process+ProcessFailed: Exception of type 'Fsdk.Process+ProcessFailed' was thrown.
   at Fsdk.Process.ProcessResult.Unwrap(String errMsg)
   at Fsdk.Process.ProcessResult.UnwrapDefault()
   at FSI_0002.RunPrettier(String arguments)
   at <StartupCode$FSI_0002>.$FSI_0002.main@()
Stopped due to error
Error: Process completed with exit code 1.
```

[1] https://github.com/actions/checkout/issues/1048