GitHubCI: install ca-certificates pkg explicitly
It turns out that dotnet6 package didn't really depend on
curl or git being installed, because as soon as you passed
--no-install-recommends to their installation, dotnet6
broke (but it only broke if it needed to retreive some pkg
from nuget actually). What was happening under the hood was
that when installing git or curl without the flag
--no-install-recommends, it was pulling `ca-certificates`
as a recommended package, which is needed to be able to
download nuget packages.

So in this commit we add `ca-certificates` to all .NET setup
steps explicitly, and now we can add --no-install-recommends
to the git & curl apt installation commands (except for the
commitlint tests job, which seems to have a different issue).