tsconfig: add comment
Because the compiler errors that were being prevented when[1]
this setting was introduced were:

```
commitlint/plugins.ts(95,29): error TS2802: Type 'IterableIterator<[number, string]>' can only be iterated through when using the '--downlevelIteration' flag or with a '--target' of 'es2015' or higher.
commitlint/plugins.ts(217,29): error TS2802: Type 'Set<unknown>' can only be iterated through when using the '--downlevelIteration' flag or with a '--target' of 'es2015' or higher.
commitlint/plugins.ts(223,29): error TS2802: Type 'Set<unknown>' can only be iterated through when using the '--downlevelIteration' flag or with a '--target' of 'es2015' or higher.
```

[1] d77fa45f1a8055c328e8feea1e0873ac4766596b