scripts/checkCommits1by1: fix IndexOutOfRangeEx
We tried to make this script support legit absence of
CI status via [1] but that caused IndexOutOfRangeExs.
This change now makes sure that the commitMsg is extracted
from the 1st API call (which is not dependent on a CI
status being present) so it should fix the problem (the
use of 0-th hardcoded indexing was a big smell here that
turned out to cause this problem).

[1] 8c00fed7d24589d7b70957cc872b1a1e57a76f93