commitlint.sh: rather use result of 'npm i'
It could happen that someone modifies this .sh file in
the future, changing the `npm install` to not install
the proper package(s) anymore, but npx still works. By
changing it to use ./node_modules/ we force this line to
depend on the results of the previous line (so that it
has a raison-d'etre).