scripts,FileConventions: improve prev commit
Previous commit [1] was less than ideal because:

1. It assumed that 1st "paragraph" of commit message would be
one line. Granted, if it's not, then commit message would not be
correct (because 2nd line should always be empty), but it's
commitlint the tool in charge of checking that. Let's not assume
anything in our wrap script.

2. It left some variable as unused (newLineIndex), now we remove
it here.

3. As a result of fixing (1) above, we need to use
StringSplitOptions.None instead of .RemoveEmptyEntries, so that
the String.Join doesn't join paragraphs (converting 2 joined
EOLs into a single one).

[1] 4c7790fd217c92e8f51e28aef84d056a29094674