commitlint.sh,CI: use a previous version to fix CI
The last version of commitlint is making our tests fail. My
suspicions for why are the following:

* After their massive refactoring migrating to ESM [1], they
somehow changed their sub-package @commitlint/types, which I
reported as bug 3942 [2].

* When fixing bug 3942 [3], they updated their dependency to
`@types/conventional-commits-parser` to a higher version (from
3.0.6 to 5.0.0). Version 5.0.0 of that package is about 18
days old, however, version 5.0.0 of package
`conventional-commits-parser` [4] is 6 months old! But the
worst is that the previous version 3.0.6 (which is the one
that commitlint was using before this update) was published
about 4 years ago! So I'm guessing our plugins were compatible
with very old versions of this internal component and now its
update has exposed bugs.

[1] https://github.com/conventional-changelog/commitlint/commit/342373559bdf7c783c4ef37ff05dc38a5f681159
[2] https://github.com/conventional-changelog/commitlint/issues/3942
[3] https://github.com/conventional-changelog/commitlint/commit/5a01f59661f0b908802728389631965eb8b49d47
[4] https://www.npmjs.com/package/conventional-commits-parser